changing pid type

+++ src/main.rs
@@ -39,7 +39,7 @@ struct Opt {
 #[derive(Debug, PartialEq, Eq)]
 pub struct FindIitRow {
     pub datid: u32,
-    pub pid: i64,
+    pub pid: i32,
     pub query: String,
     pub backend_start: DateTime<Local>,
     pub xact_start: DateTime<Local>,
@@ -57,7 +57,7 @@ async fn find_iit(client: &Client) -> Result<Vec<FindIitRow>, Box<dyn std::error
         "SELECT datid,pid,query,backend_start,xact_start,query_start,state_change from pg_stat_activity where state = 'idle in transaction' order by state_change";
     for row in client.query(query_str, &[]).await? {
         let datid: u32 = row.get(0);
-        let pid: i64 = row.get(1);
+        let pid: i32 = row.get(1);
         let query: &str = row.get(2);
         let backend_start: DateTime<Local> = row.get(3);
         let xact_start: DateTime<Local> = row.get(4);
